DeviceAdapter compiler warning removal
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17194 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Mightex_C_Cam/Mightex_USBCamera.cpp
@@ -343,7 +343,7 @@ void FrameCallBack(TProcessedDataProperty *Attributes, unsigned char *BytePtr)
 									tmpWord = 255;
 								//for(int j = 0; j < 3; j++)
 								{
-									*p = tmpWord;
+									*p = (BYTE) tmpWord;
 									p++;
 								}
 							}
@@ -351,7 +351,7 @@ void FrameCallBack(TProcessedDataProperty *Attributes, unsigned char *BytePtr)
 							{// 12 bit
 								if (tmpWord >= 0xFFF)
 									tmpWord = 0xFFF;
-								*p = tmpWord >> 4;
+								*p = (BYTE) (tmpWord >> 4);
 								p++;
 								*p = tmpWord & 0x0F;
 								p++;
